server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.58.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
itsdangerous==2.2.0
